chore: refactor text caser out of schema generator
